requirements.txt
webdriver-manager==3.2.2
selenium
python-dotenv
